// src/components/gameArt.generated.ts
export const cardArt: Record<string, string> = {
  "1-K-01": "/pygem/images/noir1-1.png",
  "1-K-02": "/pygem/images/noir1-1.png",
  "1-K-03": "/pygem/images/noir1-2.png",
  "1-K-04": "/pygem/images/noir1-1.png",
  "1-K-05": "/pygem/images/noir1-2.png",
  "1-K-06": "/pygem/images/noir1-1.png",
  "1-K-07": "/pygem/images/noir1-2.png",
  "1-K-08": "/pygem/images/noir1-2.png",
  "1-W-01": "/pygem/images/white1-1.png",
  "1-W-02": "/pygem/images/white1-1.png",
  "1-W-03": "/pygem/images/white1-2.png",
  "1-W-04": "/pygem/images/white1-2.png",
  "1-W-05": "/pygem/images/white1-2.png",
  "1-W-06": "/pygem/images/white1-2.png",
  "1-W-07": "/pygem/images/white1-2.png",
  "1-W-08": "/pygem/images/white1-2.png",
  "1-U-01": "/pygem/images/blue1-1.png",
  "1-U-02": "/pygem/images/blue1-2.png",
  "1-U-03": "/pygem/images/blue1-2.png",
  "1-U-04": "/pygem/images/blue1-1.png",
  "1-U-05": "/pygem/images/blue1-2.png",
  "1-U-06": "/pygem/images/blue1-1.png",
  "1-U-07": "/pygem/images/blue1-2.png",
  "1-U-08": "/pygem/images/blue1-2.png",
  "1-G-01": "/pygem/images/green1-1.png",
  "1-G-02": "/pygem/images/green1-2.png",
  "1-G-03": "/pygem/images/green1-2.png",
  "1-G-04": "/pygem/images/green1-1.png",
  "1-G-05": "/pygem/images/green1-1.png",
  "1-G-06": "/pygem/images/green1-2.png",
  "1-G-07": "/pygem/images/green1-2.png",
  "1-G-08": "/pygem/images/green1-2.png",
  "1-R-01": "/pygem/images/red1-1.png",
  "1-R-02": "/pygem/images/red1-1.png",
  "1-R-03": "/pygem/images/red1-1.png",
  "1-R-04": "/pygem/images/red1-2.png",
  "1-R-05": "/pygem/images/red1-1.png",
  "1-R-06": "/pygem/images/red1-1.png",
  "1-R-07": "/pygem/images/red1-2.png",
  "1-R-08": "/pygem/images/red1-1.png",
  "2-K-01": "/pygem/images/noir2-2.png",
  "2-K-02": "/pygem/images/noir2-1.png",
  "2-K-03": "/pygem/images/noir2-2.png",
  "2-K-04": "/pygem/images/noir2-2.png",
  "2-K-05": "/pygem/images/noir2-1.png",
  "2-K-06": "/pygem/images/noir2-2.png",
  "2-W-01": "/pygem/images/white2-2.png",
  "2-W-02": "/pygem/images/white2-1.png",
  "2-W-03": "/pygem/images/white2-2.png",
  "2-W-04": "/pygem/images/white2-1.png",
  "2-W-05": "/pygem/images/white2-2.png",
  "2-W-06": "/pygem/images/white2-2.png",
  "2-U-01": "/pygem/images/blue2-1.png",
  "2-U-02": "/pygem/images/blue2-2.png",
  "2-U-03": "/pygem/images/blue2-2.png",
  "2-U-04": "/pygem/images/blue2-1.png",
  "2-U-05": "/pygem/images/blue2-1.png",
  "2-U-06": "/pygem/images/blue2-2.png",
  "2-G-01": "/pygem/images/green2-2.png",
  "2-G-02": "/pygem/images/green2-2.png",
  "2-G-03": "/pygem/images/green2-2.png",
  "2-G-04": "/pygem/images/green2-2.png",
  "2-G-05": "/pygem/images/green2-2.png",
  "2-G-06": "/pygem/images/green2-1.png",
  "2-R-01": "/pygem/images/red2-1.png",
  "2-R-02": "/pygem/images/red2-2.png",
  "2-R-03": "/pygem/images/red2-2.png",
  "2-R-04": "/pygem/images/red2-2.png",
  "2-R-05": "/pygem/images/red2-1.png",
  "2-R-06": "/pygem/images/red2-1.png",
  "3-K-01": "/pygem/images/noir3-2.png",
  "3-K-02": "/pygem/images/noir3-2.png",
  "3-K-03": "/pygem/images/noir3-2.png",
  "3-K-04": "/pygem/images/noir3-2.png",
  "3-W-01": "/pygem/images/white3-2.png",
  "3-W-02": "/pygem/images/white3-1.png",
  "3-W-03": "/pygem/images/white3-1.png",
  "3-W-04": "/pygem/images/white3-1.png",
  "3-U-01": "/pygem/images/blue3-2.png",
  "3-U-02": "/pygem/images/blue3-1.png",
  "3-U-03": "/pygem/images/blue3-2.png",
  "3-U-04": "/pygem/images/blue3-1.png",
  "3-G-01": "/pygem/images/green3-1.png",
  "3-G-02": "/pygem/images/green3-2.png",
  "3-G-03": "/pygem/images/green3-2.png",
  "3-G-04": "/pygem/images/green3-2.png",
  "3-R-01": "/pygem/images/red3-1.png",
  "3-R-02": "/pygem/images/red3-1.png",
  "3-R-03": "/pygem/images/red3-1.png",
  "3-R-04": "/pygem/images/red3-1.png"
};
export const nobleArt: Record<string, string> = {
  "N-01": "/pygem/images/green-noble-1.png",
  "N-02": "/pygem/images/red-noble-1.png",
  "N-03": "/pygem/images/white-noble-2.png",
  "N-04": "/pygem/images/blue-noble-1.png",
  "N-05": "/pygem/images/red-noble-2.png",
  "N-06": "/custom-nobles/N-06.PNG",
  "N-07": "/custom-nobles/N-07.PNG",
  "N-08": "/pygem/images/white-noble-1.png",
  "N-09": "/pygem/images/noir-noble-2.png",
  "N-10": "/pygem/images/blue-noble-2.png"
};
export const deckArt: Record<number, string> = {
  "1": "/pygem/images/level-1-deck.png",
  "2": "/pygem/images/level-2-deck.png",
  "3": "/pygem/images/level-3-deck.png"
};
export const gemArt: Record<string, string> = {
  "white": "/pygem/icons/diamond-blue-inline.svg",
  "blue": "/pygem/icons/sapphire-octagon-inline.svg",
  "green": "/pygem/icons/emerald-square-inline.svg",
  "red": "/pygem/icons/ruby-red-diamond-inline.svg",
  "black": "/pygem/icons/obsidian-rectangle-inline.svg",
  "gold": "/pygem/icons/gold-coin-inline.svg"
};